+++ CHANGELOG.md
@@ -10,6 +10,9 @@ Full documentation for rocSOLVER is available at [rocsolver.readthedocs.io](http
 - Generalized symmetric- and hermitian-definite partial eigensolvers:
     - SYGVX (with batched and strided\_batched versions)
     - HEGVX (with batched and strided\_batched versions)
+- Eigensolver routines for symmetric/hermitian matrices using Jacobi algorithm:
+    - SYEVJ (with batched and strided\_batched versions)
+    - HEEVJ (with batched and strided\_batched versions)
 - Added --profile_kernels option to rocsolver-bench, which will include kernel calls in the
   profile log (if profile logging is enabled with --profile).